+++ Ruijie-WPF/Controls/MyIconButton.xaml.vb
@@ -183,17 +183,15 @@ End Class
 Partial Public Module ModAnimation
     Public Sub AniDispose(Control As MyIconButton, RemoveFromChildren As Boolean, Optional CallBack As ParameterizedThreadStart = Nothing)
         If Not Control.IsHitTestVisible Then Return
-        Control.IsHitTestVisible = False
-        AniStart({
-                 AaScaleTransform(Control, -1.5, 200,, New AniEaseInFluent),
-                 AaCode(Sub()
-                            If RemoveFromChildren Then
-                                CType(Control.Parent, Object).Children.Remove(Control)
-                            Else
-                                Control.Visibility = Visibility.Collapsed
-                            End If
-                            If CallBack IsNot Nothing Then CallBack(Control)
-                        End Sub,, True)
-        }, "MyIconButton Dispose " & Control.Uuid)
+        AniDisposeCore(Control,
+            Sub()
+                If RemoveFromChildren Then
+                    RemoveFromParent(Control)
+                Else
+                    Control.Visibility = Visibility.Collapsed
+                End If
+                If CallBack IsNot Nothing Then CallBack(Control)
+            End Sub,
+            "MyIconButton Dispose")
     End Sub
 End Module